requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
reportlab==4.0.9
python-multipart==0.0.6
pytest==7.4.4
httpx==0.26.0
pytesseract==0.3.10
Pillow==10.2.0
